pyproject.toml
[tool.poetry]
name = "tg-bot"
version = "0.1.0"
description = ""
authors = ["MonkeyBusiness <crimson.oliver@gmail.com>"]
readme = "README.md"
packages = [{include = "tg_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
aiogram = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
